pyproject.toml
[project]
name = "bitwig-mcp-server"
version = "0.0.1"
description = "MCP server for Bitwig Studio."
authors = [{ name = "John Stanford", email = "787382+jxstanford@users.noreply.github.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "mcp[cli]>=1.4.1",
    "pydantic>=2.10.2",
    "pydantic-settings>=2.6.1",
    "pytest-asyncio>=0.25.3",
    "python-osc>=1.8.3",
    "chromadb>=0.4.18",
    "sentence-transformers>=2.2.2",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://jxstanford.github.io/bitwig-mcp-server/"
Repository = "https://github.com/jxstanford/bitwig-mcp-server"
Documentation = "https://jxstanford.github.io/bitwig-mcp-server/"

[project.scripts]
bitwig-browser-index = "bitwig_mcp_server.utils.index_browser:main"
bitwig-device-recommend = "bitwig_mcp_server.utils.recommend_devices:main"
bitwig-enhance-index = "bitwig_mcp_server.utils.enhance_index:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.20.0",
    "mypy>=0.991",
    "pytest-cov>=4.0.0",
    "ruff>=0.6.9",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "mcp[cli]>=1.4.1",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["bitwig_mcp_server"]

[tool.mypy]
files = ["bitwig_mcp_server"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # BroadException
    "TRY003",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["bitwig_mcp_server"]
